+++ packages/keyring-snap-bridge/src/index.ts
@@ -2,5 +2,4 @@ export * from './options';
 export * from './types';
 export * from './SnapKeyring';
 export * from './SnapKeyringV1';
-export * from './SnapKeyringV2';
 export type * from './SnapKeyringMessenger';