+++ packages/keyring-eth-simple/CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `./v2` subpath export for keyring v2 implementation ([#513](https://github.com/MetaMask/accounts/pull/513))
+  - `SimpleKeyring` and `SimpleKeyringOptions` are now available from `@metamask/eth-simple-keyring/v2`.
+
+### Changed
+
+- **BREAKING:** Move and rename `SimpleKeyringV2` and `SimpleKeyringV2Options` to the new `./v2` subpath export ([#513](https://github.com/MetaMask/accounts/pull/513))
+  - `SimpleKeyringV2` is now `SimpleKeyring` from `@metamask/eth-simple-keyring/v2`.
+  - `SimpleKeyringV2Options` is now `SimpleKeyringOptions` from `@metamask/eth-simple-keyring/v2`.
+
 ## [11.1.2]
 
 ### Changed